FAAC SLH/Spa 20-bits counter

+++ lib/subghz/protocols/faac_slh.h
@@ -42,7 +42,7 @@ bool subghz_protocol_faac_slh_create_data(
     FlipperFormat* flipper_format,
     uint32_t serial,
     uint8_t btn,
-    uint16_t cnt,
+    uint32_t cnt,
     uint32_t seed,
     const char* manufacture_name,
     uint32_t frequency,
